requirements.txt
Flask==3.0.3
flask-restful
mysql-connector-python
psycopg-binary
passlib
flask-jwt-extended
email-validator
serverless-wsgi
spotipy
werkzeug